Added Custom Named Hooks and Handles which imported in App.js
