+++ src/handlers/add-watch.handle.js
@@ -1,0 +1,3 @@
+export function addWatchedHandler(movie, setWatched) {
+  setWatched((watched) => [...watched, movie]);
+}
