+++ src/handlers/close-movie.handle.js
@@ -1,0 +1,3 @@
+export function closeMovieHandler(setSelectedId) {
+  setSelectedId(null);
+}
