+++ src/handlers/delete-watched.handle.js
@@ -1,0 +1,3 @@
+export function deleteWatchedHandler(id, setWatched) {
+  setWatched((watched) => watched.filter((movie) => movie.imdbID !== id));
+}
