+++ src/handlers/select-movie.handle.js
@@ -1,0 +1,3 @@
+export function selectMovieHandler(id, setSelectedId) {
+  setSelectedId((selectedId) => (id === selectedId ? null : id));
+}
